+++ src/cmd/grit/interactive.go
@@ -13,7 +13,6 @@ import (
 	"gopkg.in/src-d/go-git.v4/plumbing/transport"
 
 	"github.com/jmalloc/grit/src/grit"
-	"github.com/jmalloc/grit/src/grit/pathutil"
 	"github.com/urfave/cli"
 )
 
@@ -77,17 +76,10 @@ func choose(c *cli.Context, opt []string) (int, bool) {
 }
 
 func chooseCloneDir(cfg grit.Config, c *cli.Context, dirs []string) (string, bool) {
-	gosrc, _ := pathutil.GoSrc()
 	var opts []string
 
 	for _, dir := range dirs {
-		if rel, ok := pathutil.RelChild(gosrc, dir); ok && gosrc != "" {
-			opts = append(opts, fmt.Sprintf("[go] %s", rel))
-		} else if rel, ok := pathutil.RelChild(cfg.Clone.Root, dir); ok {
-			opts = append(opts, fmt.Sprintf("[grit] %s", rel))
-		} else {
-			opts = append(opts, dir)
-		}
+		opts = append(opts, formatDir(cfg, dir))
 	}
 
 	if i, ok := choose(c, opts); ok {